Replace old render crate, fix brush lightmaps, enable msaa
