+++ Cargo.toml
@@ -25,7 +25,7 @@ regex = "0.2.6"
 rodio = "0.11.0"
 shaderc = "0.6.2"
 # wgpu = "0.6.0"
-wgpu = { git = "https://github.com/gfx-rs/wgpu-rs", branch = "master" }
+wgpu = { git = "https://github.com/gfx-rs/wgpu-rs", rev = "d272a69" }
 
 # gfx pre-ll deps
 #
